move fitRouteInCam into new file camUtils.ts
